Refactor the code and update the readme
